Server side implementation and more app side work
The server side is now implemented. The createEntry is reformed using
the new form object values, which are then inserted into the Entry
table.

App side, a case was added which catches if a user inputs other and
then deletes it. Without the case, it throws an error as incomplete.