Catch case where a user might select NONE and OTHER
To catch the case where a user might select NONE and attempt to write
in OTHER, I simply made it so that when a user selects NONE, the entry
box for OTHER disappears completely. As soon as NONE is toggled again
in the event they made a mistake, the entry box reappears.