Adds data collection to the selfie questions
Data is now collected for the selfie questions in the global array
items. Items is an array of photos (each photo being called "item").
How many photos are appended to the array depends on the numSelfie
input. It then pushes the "item" (denoting a photo) with the following
JSON structure:
items = [
	{
	    //The number of the photo to be displayed for the user
	    //"Photo #1", "Photo #2", etc.
	    "name": "Photo #" + (i+1),

	    //The responses to the categorizations, initialized as
	    //an empty array
	    description: [],

	    //The "other" field value for the categories
	    otherSelfie: "",

	    //The 4 Likert-style questions where users select 1-6
	    boredSelfie: "",
	    documentSelfie: "",
	    reactionSelfie:"",
	    communicateSelfie: ""
    	}

	//OTHER ITEMS WILL BE HERE IF THEY SELECT MORE THAN 1
	]